% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtheme.R
\name{theme_q}
\alias{theme_q}
\title{A simple-looking ggplot2 theme for publication: some adjustments from the theme_par function.}
\usage{
theme_q(..., bg = "white")
}
\arguments{
\item{...}{parameters passed to theme_par function}

\item{bg}{color for background}
}
\value{
ggplot2 theme object
}
\description{
A simple-looking ggplot2 theme for publication: some adjustments from the theme_par function.
}
\examples{
library(ggpubr)
data.frame(x=1:10,y=2:11)\%>\%
  ggscatter(x="x",y="y",color = "x")+
  theme_q()
}
\author{
qinti
}
