% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtheme.R
\name{theme_t}
\alias{theme_t}
\title{A simple-looking ggplot2 theme for publication: some adjustments from the theme_pubr function.}
\usage{
theme_t(..., bg = "white")
}
\arguments{
\item{...}{parameters passed to theme_pubr function}

\item{bg}{color for background}
}
\value{
ggplot2 theme object
}
\description{
A simple-looking ggplot2 theme for publication: some adjustments from the theme_pubr function.
}
\examples{
library(ggpubr)
data.frame(x = 1:10, y = 2:11) \%>\%
  ggscatter(x = "x", y = "y", color = "x") +
  theme_t()
}
\author{
qinti
}
