% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtheme.R
\name{ttsave}
\alias{ttsave}
\title{A funtion to save ggplot project in fixed style}
\usage{
ttsave(name, w, h)
}
\arguments{
\item{name}{filename of pdf or png}

\item{w}{the width of plot}

\item{h}{the height of plot}
}
\value{
No return value. This function is called for its side effects.
}
\description{
A funtion to save ggplot project in fixed style
}
\examples{
library(ggplot2)
data <- data.frame(x = 1:10, y = 2:11)
ggplot(data, mapping = aes(x = x, y = y)) +
  geom_point()
tmpfile <- tempfile(fileext = ".pdf")
ttsave(tmpfile, 5.5, 5.5)
}
\author{
qinti
}
